% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{added_seats}
\alias{added_seats}
\title{Beregner Sainte-Laguës med modificering}
\usage{
added_seats(candidates, votes, seats, attained, by)
}
\arguments{
\item{candidates}{vektor med kandidater / partier}

\item{votes}{vektor med deres respektive stemmetal}

\item{seats}{antal mandater i alt}

\item{attained}{hvor mange de i forvejen har fået}

\item{by}{hvad devisoren skal hoppe med}
}
\description{
Beregner Sainte-Laguës med modificering
}
