% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{beregn_endeligt}
\alias{beregn_endeligt}
\title{Beregner det endelige resultat og fordeler tillægsmandater på storkredse}
\usage{
beregn_endeligt(kredsmandater_beregnet, tillaeg_landsdel_beregnet)
}
\arguments{
\item{kredsmandater_beregnet}{output fra beregn_kredsmandater()}

\item{tillaeg_landsdel_beregnet}{output fra beregn_tillaeg_landsdel()}
}
\description{
Beregner det endelige resultat og fordeler tillægsmandater på storkredse
}
