% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{beregn_kredsmandater}
\alias{beregn_kredsmandater}
\title{Beregner kredsmandater}
\usage{
beregn_kredsmandater(storkreds_votes, kredsmandater)
}
\arguments{
\item{storkreds_votes}{stemmer fordelt på storkredse}

\item{kredsmandater}{oversigt over antallet af kredsmandater}
}
\description{
Beregner kredsmandater
}
