% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{beregn_mandater}
\alias{beregn_mandater}
\title{Beregner fordelingen af mandater}
\usage{
beregn_mandater(storkreds_votes)
}
\arguments{
\item{storkreds_votes}{stemmer fordelt på storkredse}
}
\description{
Beregner fordelingen af mandater
}
