% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{beregn_manglende}
\alias{beregn_manglende}
\title{Beregner manglende tillægsmandater}
\usage{
beregn_manglende(kredsmandater_beregnet)
}
\arguments{
\item{kredsmandater_beregnet}{output fra beregn_kredsmandater()}
}
\description{
Beregner manglende tillægsmandater
}
