% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nye_mandater.R
\name{beregn_nye_stemmer}
\alias{beregn_nye_stemmer}
\title{Beregn nye stemmer p책 baggrund af meningsm책ling}
\usage{
beregn_nye_stemmer(storkreds_res, pp)
}
\arguments{
\item{storkreds_res}{stemmerne fra FV15}

\item{pp}{ny meningsm책ling}
}
\value{
en tibble
}
\description{
Funktionen tager stemmerne fra FV15 og omfordeler dem efter meningsm책linger.
}
