% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{beregn_tillaeg_landsdel}
\alias{beregn_tillaeg_landsdel}
\title{Beregner fordelingen af tillægsmandater på partier og landsdele}
\usage{
beregn_tillaeg_landsdel(kredsmandater_beregnet, manglende_beregnet)
}
\arguments{
\item{kredsmandater_beregnet}{output fra beregn_kredsmandater()}

\item{manglende_beregnet}{output fra beregn_manglende()}
}
\description{
Beregner fordelingen af tillægsmandater på partier og landsdele
}
