% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dHondt_df}
\alias{dHondt_df}
\title{Beregner dHondt}
\usage{
dHondt_df(party, votes, seats)
}
\arguments{
\item{party}{en vektor med partier i valget}

\item{votes}{en vektor med partiernes stemmer}

\item{seats}{integer med antal mandater i alt}
}
\description{
Beregner dHondt
}
