% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fjern_under_graensen}
\alias{fjern_under_graensen}
\title{Fjerner partier under spærregrænsen}
\usage{
fjern_under_graensen(storkreds_votes)
}
\arguments{
\item{storkreds_votes}{stemmer fordelt på storkredse}
}
\description{
Fjerner partier under spærregrænsen
}
