% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Runder men beholder summen}
\usage{
round_preserve_sum(x, digits = 0)
}
\arguments{
\item{x}{tal der skal rundes}

\item{digits}{antal decimaler}
}
\description{
Runder men beholder summen
}
